fix: support reverse runner hot offload selection
